feature-commit: refactor: chat, models and base class
- Updated OllamaModel and Phi2Model constructors to ensure proper parameter passing for model initialization.
- Changed import path for create_chat_agent from workflows.chat_agent to workflows.nodes.chat_agent.
- Removed commented-out tool functions and related code in chat.py to streamline the file.
